added bidirectional version
backward:
1) forward propagation:
a. if collision number exceeds limit, then backtrace
b. if not near the waypoint then count as invalid
c. if in terms of funnel distance, the endpoint is too far then count as invalid

2). backward propagation:
1) forward propagation:
a. if collision number exceeds limit, then backtrace
b. if not near the waypoint then count as invalid